% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_filter.R
\name{envelope_filter}
\alias{envelope_filter}
\title{Apply an envelope model for a set of predictors for multiple taxa}
\usage{
envelope_filter(
  taxa,
  predictors,
  vars = elements::VariableNames,
  limit = "min_max"
)
}
\arguments{
\item{taxa}{A vector of strings containing one or more taxon codes. See \code{elements::ModelledTaxaCodes}. Optional}

\item{predictors}{A data frame of predictors. Must include one or more of the following columns: L, M, N, R, S, SD, GP, tmax_sm, tmin_wt, prec_wt, and prec_sm}

\item{vars}{A vector of strings, containing one or more variables as present in \code{elements::VariableNames}, which must also be present in the predictors data frame.}

\item{limit}{A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q10_q90", "q25_q75".}
}
\value{
The predictors data frame containing an additional column 'within_limit' indicating whether each row of predictord is wthin the specified limits.
}
\description{
Append a column to set of predictors which indicates whether the predictors
in each row are within a set of distributional limits
for a particular taxon and set of variables.
}
\details{
This function acts as an envelope, allowing the screening of large quantities
of predictor data for the suitability of multiple taxa before applying the more
computationally expensive SVM models in \code{elements::predict_occ} and \code{elements::predict_occ_taxon}.
}
\examples{
\dontrun{
elements::envelope_filter_taxon(taxa = NULL, predictors = elements::ExampleData2)
}
}
