% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/envelope_fillter_taxon.R
\name{envelope_filter_taxon}
\alias{envelope_filter_taxon}
\title{Apply an envelope model for a set of predictors for one taxon}
\usage{
envelope_filter_taxon(
  taxon,
  predictors,
  vars = elements::VariableNames,
  limit = "min_max"
)
}
\arguments{
\item{taxon}{The taxon_code, see \code{elements::TaxonomicBackbone}.}

\item{predictors}{A data frame of predictors. Must include one or more of the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, and bio17}

\item{vars}{A vector of strings, containing one or more variables as present in \code{elements::VariableNames}, which must also be present in the predictors data frame.}

\item{limit}{A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q10_q90", "q25_q75".}
}
\value{
The predictors data frame containing an additional column 'within_limit' indicating whether each row of predictord is wthin the specified limits.
}
\description{
Append a column to set of predictors which indicates whether the predictors
in each row are within a set of distributional limits
for a particular taxon and set of variables.
}
\details{
This function acts as an envelope, allowing the screening of large quantities
of predictor data for the suitability of a taxon before applying the more
computationally expensive SVM models in \code{elements::predict_occ} function
and to a lesser extent the euclidean distance \code{elements::calc_distance} function.
}
\examples{
\dontrun{
elements::envelope_filter(taxon = "stellaria_graminea", predictors = elements::ExampleData1)
}
}
