% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EuroSL}
\alias{EuroSL}
\title{EuroSL data}
\format{
A data frame with 53643 rows and 6 columns, the definitions of which are:
\describe{
\item{TaxonUsageID}{The unique code identifying the taxon.}
\item{TaxonName}{The taxon name.}
\item{NameAuthor}{The author who described the taxon concept.}
\item{TaxonRank}{The rank of the taxon.}
\item{IsChildTaxonOfID}{The immediate parent taxon.}
\item{IsChildTaxonOf}{The unique code identifying the parent taxon.}
}
}
\usage{
EuroSL
}
\description{
Selected columns for the accepted taxa in the European European Species List (EuroSL) database.
}
\details{
\code{EuroSL}
}
\keyword{datasets}
