% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData}
\alias{ExampleData}
\title{Example predictor data}
\format{
A data frame with 200 rows and 12 columns, the definitions of which are:
\describe{
\item{Presence}{A column of type string, containing either "Present" or "Absent".}
}
}
\usage{
ExampleData
}
\description{
A dataset containing a randomised sample of 100 presences and 100 absences from the training and test data for the Stellaria graminea (stellaria_graminea)
}
\details{
\code{ExampleData}
}
\keyword{datasets}
