% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExamplePlots}
\alias{ExamplePlots}
\title{A set of example vegetation plots}
\format{
A data frame with 93 rows and 4 columns, the definitions of which are:
\describe{
\item{plot_id}{The plot id.}
\item{taxon}{The scientific names of the taxa present in the example plot.}
\item{taxon_code}{The taxon, see \code{elements::ModelledTaxaCodes}.}
\item{cover_perc}{The percentage cover of the taxa present in example plot.}
}
}
\usage{
ExamplePlots
}
\description{
A dataset containing the taxon presences and percentage cover for a set of vegetation plots recorded in a spring-fed calcareous fen - Redcar Field, County Durham, United Kingdom.
These plots are not currently included in the EVA and were not included in the model training data, therefore representing an independent sample.
These plots are used to form the example scenarios present in \code{elements::ExampleScenarios}.
}
\details{
\code{ExamplePlots}
}
\keyword{datasets}
