% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleScenarios}
\alias{ExampleScenarios}
\title{Predictor data for three example scenarios}
\format{
A data frame with 48 rows and 16 columns, the definitions of which are:
\describe{
\item{L}{Light}
\item{M}{Soil Moisture}
\item{N}{Soil Nitrogen}
\item{R}{Reaction}
\item{S}{Salinity}
\item{SD}{Soil Disturbance}
\item{GP}{Grazing Pressure}
\item{tmax_sm}{Maximum temperature in the warmest month}
\item{tmin_wt}{Minimum temperature in the coldest month}
\item{prec_wt}{Precipitation in the wettest quarter}
\item{prec_sm}{Precipitation in the driest quarter}
\item{scenario_code}{The scenario code.}
\item{scenario_stage}{The scenario stage, either 1) baseline, or 2) projection.}
\item{climate_scenario}{The SSP climate scenario.}
\item{timeslice}{The scenario timeslice, either a year or period.}
\item{scenario_name}{The scenario name.}
}
}
\usage{
ExampleScenarios
}
\description{
A dataset containing predictor data for four sets of example scenarios:
(A) Climate Change: a.i - ssp126, a.ii - ssp245, a.iii - ssp370, a.iv - ssp585.
(B) Grazing: b.i - Grazing intensification, b.i - Grazing reduction
(C) Nutrient: c.i - Nutrient Enrichment, c.ii - Nutrient Reduction
(D) Interactions: d.i - Climate Change + Nutrient Reduction + Grazing Reduction,
d.ii - Climate Change + Nutrient Addition + Grazing Reduction,
d.iii - Climate Change + Nutrient Reduction + Grazing Intensification,
d.iv - Climate Change + Nutrient Addition + Grazing Intensification
}
\details{
\code{ExampleScenarios}
}
\keyword{datasets}
