% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extract_svs.R
\name{extract_svs}
\alias{extract_svs}
\title{Extract the support vectors from an SVM model}
\usage{
extract_svs(taxon)
}
\arguments{
\item{taxon}{A string containing the taxon code for a taxon, see \code{elements::TaxonomicBackbone}.}
}
\value{
A data frame containing 11 columns, containing the support vector points for each variable in \code{length(elements::VariableNames)}
}
\description{
Extract the support vectors from the SVM model for a taxon
and descale them.
}
\examples{
\dontrun{
elements::startup(); elements::extract_svs(taxon = "carex_nigra")
}
}
