% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{featureImportance}
\alias{featureImportance}
\title{Feature importance}
\format{
A data frame with 528 rows and 4 columns, the definitions of which are:
\describe{
\item{species}{The species, represented by a UKSI TVK.}
\item{variable}{A string representing the names of the model variable (), baseline (\emph{baseline}), and full model (\emph{full_model})}
\item{permutation}{The \code{DALEX::model_parts} permuation}
\item{dropout_loss}{The dropout loss associated with a particular variable and permuation.}
}
}
\usage{
featureImportance
}
\description{
A dataset containing the feature importance of each model variable value for each species model
generated using the \code{DALEX::model_parts} function.
}
\details{
\code{featureImportance}
}
\keyword{datasets}
