% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_parent_taxa.R
\name{get_parent_taxa}
\alias{get_parent_taxa}
\title{Retrieve the parent taxa for a given taxon.}
\usage{
get_parent_taxa(taxon)
}
\arguments{
\item{taxon}{The name of a Taxon in the "TaxonName" column of the \code{elements::EuroSL} data frame, or the "taxon_name" column of the \code{elements::TaxonomicBackbone} data frame.}
}
\value{
A data frame in wide format containing the parent taxa for the specified taxon.
}
\description{
Retrieve the parent taxa for a given taxon.
}
\examples{
elements::get_parent_taxa(taxon = "Populus nigra subsp. betulifolia")
}
