% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Gradients}
\alias{Gradients}
\title{Univariate gradient intervals for each environmental variable}
\format{
A named list with \code{length(elements::Gradients)} items.
\describe{
\item{L}{Light - 0:10}
\item{M}{Soil Moisture - 0:10}
\item{N}{Soil Nitrogen - 0:10}
\item{R}{Reaction - 0:10}
\item{S}{Salinity - 0:10}
\item{SD}{Soil Disturbance - 0:1}
\item{GP}{Grazing Pressure - 0:1}
\item{tmax_sm}{Maximum temperature in the warmest month - -5:55}
\item{tmin_wt}{Minimum temperature in the coldest month - -30:15}
\item{prec_wt}{Precipitation in the wettest quarter - 0:1700}
\item{prec_sm}{Precipitation in the driest quarter - 0:800}
}
}
\usage{
Gradients
}
\description{
A named list containing univariate gradient intervals for each environmental variable in \code{elements::VariableNames}.
}
\details{
\code{Gradients}
}
\keyword{datasets}
