% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{HOAData}
\alias{HOAData}
\title{HOA data}
\format{
A data frame with 6604476 rows and 4 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon, see \code{elements::TaxonomicBackbone}.}
\item{x}{The variable value.}
\item{y}{The PDP value.}
\item{variable}{The variable name, see \code{elements::VariableNames}.}
}
}
\usage{
HOAData
}
\description{
A dataset containing the a set of predictions for each variable along its gradient (see \code{elements::Gradients})
produced using the \code{elements::predict_occ} function with the 'holdopt' argument set to all other variables.
}
\details{
\code{HOAData}
}
\keyword{datasets}
