% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{marginalEffects}
\alias{marginalEffects}
\title{Marginal effects plot data}
\format{
A data frame with 11586 rows and 5 columns, the definitions of which are:
\describe{
\item{species}{The species, represented by a UKSI TVK.}
\item{variable}{The model variable name ().}
\item{x}{The model vairable value.}
\item{y}{The model response value.}
\item{type}{The marginal effects plot type.}
}
}
\usage{
marginalEffects
}
\description{
A dataset containing the marginal effects data produced using the \code{DALEX::model_profile} function.
Three marginal effect types are provided: Accumulated Local (AL), Partial Dependence (PD), and Local Dependence (LD).
}
\details{
\code{marginalEffects}
}
\keyword{datasets}
