% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{MaximumProbabilities}
\alias{MaximumProbabilities}
\title{Maximum Probabilities}
\format{
A data frame with 8897 rows and 2 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon, see \code{elements::ModelledTaxaCodes}.}
\item{Present}{The maxmimum probability of presence.}
}
}
\usage{
MaximumProbabilities
}
\description{
A data frame containing the maximum predicted probability values for each taxon using the training data as predictors.
As SVM models were used there is limited embodied rarity and so these values are usally close to 1.
}
\details{
\code{MaximumProbabilities}
}
\keyword{datasets}
