% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModellingTaxaLookup}
\alias{ModellingTaxaLookup}
\title{Modelling Taxa Lookup}
\format{
A data frame with 14315 rows and 3 columns, the definitions of which are:
\describe{
\item{taxon}{The taxon name as present in the EVA.}
\item{taxon_name}{The aggregated taxon name for use in \code{elements}.}
\item{taxon_code}{The aggregated taxon code, see \code{elements::ModelledTaxaCodes}.}
}
}
\usage{
ModellingTaxaLookup
}
\description{
A lookup between the EVA taxon names, aggregated taxon names, and taxon codes used in the \code{elements} package.
}
\details{
\code{ModellingTaxaLookup}
}
\keyword{datasets}
