% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{models}
\alias{models}
\title{Ecological Niche Models}
\format{
A named list with  elements. Element names are the species, represented by a UKSI TVK..
}
\usage{
models
}
\description{
A list of Support Vector Machine (SVM) ecological niche models for each species.
}
\details{
\code{models}
}
\keyword{datasets}
