% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModelTaxa}
\alias{ModelTaxa}
\title{Information on the modelled taxa.}
\format{
A data frame with 6145 rows and 6 columns, the definitions of which are:
\describe{
\item{taxon_name}{The full taxon name, currently as present in the "Matched concept" column in the EVA.}
\item{taxon_code}{The taxon code, equivalent to the full taxon name but all lower case and with whitespace replaced with underscores.}
\item{eurosl_name}{The taxon name as it appears in EuroSL, where there is a match.}
\item{eurosl_status}{The taxonomic status of the eurosl_name.}
\item{eurosl_id}{The ID of the eurosl_name.}
\item{n_presences_eva}{The number of presences in suitable samples from the EVA.}
}
}
\usage{
ModelTaxa
}
\description{
Information on the modelled taxa in \code{elements}, including the taxon_code values underpinning all functions and objects.
}
\details{
Note that further information will be included before v1 of the package is minted.

\code{ModelTaxa}
}
\keyword{datasets}
