% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{NicheWidthData}
\alias{NicheWidthData}
\title{Niche width data}
\format{
A data frame with 1408 rows and 12, the definitions of which are:
\describe{
\item{variable}{The variable name.}
\item{mean}{The mean variable value.}
\item{min}{The minimum variable value.}
\item{max}{The maximum variable value.}
\item{median}{The median variable value.}
\item{q01}{The 1\% quantile value.}
\item{q05}{The 5\% quantile value.}
\item{q25}{The 25\% quantile value.}
\item{q75}{The 75\% quantile value.}
\item{q95}{The 95\% quantile value.}
\item{q99}{The 99\% quantile value.}
\item{taxon_code}{The taxon, see \code{elements::TaxaBackbone}.}
}
}
\usage{
NicheWidthData
}
\description{
A dataset Niche width data for the modelled taxa.
}
\details{
\code{NicheWidthData}
}
\keyword{datasets}
