% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict.R
\name{predict}
\alias{predict}
\title{Title}
\usage{
predict(species, predictors)
}
\arguments{
\item{species}{}

\item{predictors}{}
}
\value{
A data frame containing the model predictions...
}
\description{
Title
}
