% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shutdown.R
\name{shutdown}
\alias{shutdown}
\title{Remove the filehash database connection from the global environment}
\usage{
shutdown()
}
\description{
Remove the filehash database connection from the global environment
}
\examples{
elements::shutdown()
}
