% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxonomicBackbone}
\alias{TaxonomicBackbone}
\title{Taxonomic backbone}
\format{
A data frame with 6631 rows and 13 columns, the definitions of which are:
\describe{
\item{taxon_name}{The taxon names for the modelled taxa.}
\item{taxon_code}{The taxon codes used throughout the package, formed from the taxon_name values by coercing all letters to lower and replacing whitespace with underscores.}
\item{scientificName}{The full scientific name of the taxon, which includes the author. Retrieved from GBIF.}
\item{canonicalName}{The full name of the taxon. Retrieved from GBIF.}
\item{species}{The accepted name for the taxon, following GBIF.}
\item{genus}{The parent Genus taxon associated with the Species.}
\item{family}{The parent Family taxon associated with the Species.}
\item{order}{The parent Order taxon associated with the Species.}
\item{class}{The parent Class taxon associated with the Species.}
\item{phylum}{The parent Phylum taxon associated with the Species.}
\item{kingdom}{The parent Kingdom taxon associated with the Species.}
}
}
\usage{
TaxonomicBackbone
}
\description{
The taxon names and codes for the modeled taxa in the EVA, with the associated taxon concept information retireved from GBIF.
}
\details{
\code{TaxonomicBackbone}
}
\keyword{datasets}
