% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ALEData}
\alias{ALEData}
\title{ALE data}
\format{
A data frame with 2167063 rows and 4 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon, see \code{elements::ModellingTaxaLookup}.}
\item{x}{The variable value.}
\item{y}{The ALE value.}
\item{variable}{The variable name.}
}
}
\usage{
ALEData
}
\description{
A dataset containing the Accumulated Local Effect (ALE) (Molnar, 2018) data for the models, generated using 1000 samples from the training data.
}
\details{
\code{ALEData}
}
\keyword{datasets}
