% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_distance_once.R
\name{calc_distance_once}
\alias{calc_distance_once}
\title{Calculate the euclidean distance between one set of environmental predictor variables and mutliple taxa}
\usage{
calc_distance_once(
  predictors,
  taxa = ModelledTaxaCodes,
  vars = elements::VariableNames,
  method = "median",
  append = "ids"
)
}
\arguments{
\item{predictors}{A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, tmax_sm, tmin_wt, prec_wt, prec_sm, and optionally taxon}

\item{taxa}{A vector of strings containing one or more taxa to calculate distances for. See \code{elements::ModelledTaxaCodes}.}

\item{vars}{A vector of variable names, one or more of \code{elements::VariableNames}.}

\item{method}{One of "median" or "mean" indicating whether to measure the euclidean distance from the predictor variables to the median or mean values.}

\item{append}{A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.}
}
\value{
A data frame containing the euclidean distance for each taxon.
}
\description{
Calculate the euclidean distance between one set of environmental predictor variables and mutliple taxa
}
\examples{
elements::calc_distance_once(predictors = elements::ExampleScenarios[1,])
}
