% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_overlap.R
\name{calc_overlap}
\alias{calc_overlap}
\title{Calculate the niche overlap between a set of taxa}
\usage{
calc_overlap(
  taxa,
  taxa_comp = NULL,
  data = elements::HOAData,
  vars = elements::VariableNames,
  method = "D",
  average = FALSE,
  format = "wide"
)
}
\arguments{
\item{taxa}{A vector of strings containing taxon codes, see \code{elements::TaxonomicBackbone}.}

\item{taxa_comp}{A vector of string containing taxon codes, see \code{elements::TaxonomicBackbone}.}

\item{data}{A data frame containing the univariate responses of the taxa supplied in the 'taxa' and 'taxa_comp' arguments for the variables supplied in the 'vars' argument. \code{elements::HOAData} by default.}

\item{vars}{A vector of strings containing the variables for which to calculate the overlap, one or more of \code{elements::VariableNames}.}

\item{method}{A string, one of "D" or "I" indicating the overlap method (see description).}

\item{average}{A boolean (TRUE/FALSE) indicating whether to calculate the mean overlap across all variables.}

\item{format}{A boolean (TRUE/FALSE) indicating whether to return the results in wide of long format.}
}
\value{
A data frame containing the overlap values.
}
\description{
Calculate the niche overlap between all combinations of taxa as present in the
'taxa' argument, or the combinations of taxa present in the 'taxa' and 'taxa_comp'
arguments.
Two methods are curently provided:
"D" - Schoener's D (Schoener, 1968)
"I" - Warren's I (Warren et al., 2008)
}
\examples{
elements::calc_overlap(taxa = c("carex_arenaria", "carex_divulsa", "carex_hirta", "carex_sylvatica"))
}
\references{
Schoener, T.W., 1968. The Anolis Lizards of Bimini: Resource Partitioning in a Complex Fauna. Ecology 49, 704–726. https://doi.org/10.2307/1935534
Warren, D.L., Glor, R.E., Turelli, M., 2008. ENVIRONMENTAL NICHE EQUIVALENCY VERSUS CONSERVATISM: QUANTITATIVE APPROACHES TO NICHE EVOLUTION. Evolution 62, 2868–2883. https://doi.org/10.1111/j.1558-5646.2008.00482.x
}
