% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{EUNISDiagnosticTaxa}
\alias{EUNISDiagnosticTaxa}
\title{EUNISDiagnosticTaxa}
\format{
A tibble with 5122 rows and 5 columns.

\describe{
\item{eunis_code}{The EUNIS habitat code.}
\item{eunis_name}{The EUNIS habitat name.}
\item{taxon_name}{The name of the taxon as it appears in Chytrý et al (2020).}
\item{taxon_code}{The taxon, see \code{elements::ModellingTaxaLookup}.}
\item{phi_100}{The phi coefficient of association between the taxon and EUNIS habitat, a measure of fidelity.}
}
}
\usage{
EUNISDiagnosticTaxa
}
\description{
A tibble containing the diagnostic taxa associated with each EUNIS habitat.
}
\details{
\code{EUNISDiagnosticTaxa}
}
\references{
Chytrý, M., Tichý, L., Hennekens, S.M., Knollová, I., Janssen, J.A.M., Rodwell, J.S., Peterka, T., Marcenò, C., Landucci, F., Danihelka, J., Hájek, M., Dengler, J., Novák, P., Zukal, D., Jiménez-Alfaro, B., Mucina, L., Abdulhak, S., Aćić, S., Agrillo, E., Attorre, F., Bergmeier, E., Biurrun, I., Boch, S., Bölöni, J., Bonari, G., Braslavskaya, T., Bruelheide, H., Campos, J.A., Čarni, A., Casella, L., Ćuk, M., Ćušterevska, R., De Bie, E., Delbosc, P., Demina, O., Didukh, Y., Dítě, D., Dziuba, T., Ewald, J., Gavilán, R.G., Gégout, J.-C., Giusso del Galdo, G.P., Golub, V., Goncharova, N., Goral, F., Graf, U., Indreica, A., Isermann, M., Jandt, U., Jansen, F., Jansen, J., Jašková, A., Jiroušek, M., Kącki, Z., Kalníková, V., Kavgacı, A., Khanina, L., Yu. Korolyuk, A., Kozhevnikova, M., Kuzemko, A., Küzmič, F., Kuznetsov, O.L., Laiviņš, M., Lavrinenko, I., Lavrinenko, O., Lebedeva, M., Lososová, Z., Lysenko, T., Maciejewski, L., Mardari, C., Marinšek, A., Napreenko, M.G., Onyshchenko, V., Pérez-Haase, A., Pielech, R., Prokhorov, V., Rašomavičius, V., Rodríguez Rojo, M.P., Rūsiņa, S., Schrautzer, J., Šibík, J., Šilc, U., Škvorc, Ž., Smagin, V.A., Stančić, Z., Stanisci, A., Tikhonova, E., Tonteri, T., Uogintas, D., Valachovič, M., Vassilev, K., Vynokurov, D., Willner, W., Yamalov, S., Evans, D., Palitzsch Lund, M., Spyropoulou, R., Tryfon, E., Schaminée, J.H.J., 2020. EUNIS Habitat Classification: Expert system, characteristic species combinations and distribution maps of European habitats. Applied Vegetation Science 23, 648–675. https://doi.org/10.1111/avsc.12519
}
\keyword{datasets}
