% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData1}
\alias{ExampleData1}
\title{Example predictor data}
\format{
A data frame with 200 rows and 11 columns, the definitions of which are:
\describe{
\item{L}{}
\item{M}{}
\item{N}{}
\item{R}{}
\item{S}{}
\item{SD}{}
\item{GP}{}
\item{bio05}{}
\item{bio06}{}
\item{bio16}{}
\item{bio17}{}
}
}
\usage{
ExampleData1
}
\description{
A dataset containing a randomised sample of 100 presences and 100 absences from the training and test data for one taxon:
Stellaria graminea (stellaria_graminea)
}
\details{
\code{ExampleData1}
}
\keyword{datasets}
