% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleData2}
\alias{ExampleData2}
\title{Example predictor data}
\format{
A data frame with 400 rows and 12 columns, the definitions of which are:
\describe{
\item{L}{}
\item{M}{}
\item{N}{}
\item{R}{}
\item{S}{}
\item{SD}{}
\item{GP}{}
\item{bio05}{}
\item{bio06}{}
\item{bio16}{}
\item{bio17}{}
\item{taxon_code}{}
}
}
\usage{
ExampleData2
}
\description{
A dataset containing a randomised sample of 100 presences and 100 absences from the training and test data for two taxa:
Stellaria graminea (stellaria_graminea) and Silene flos-cuculi (silene_flos-cuculi).
}
\details{
\code{ExampleData2}
}
\keyword{datasets}
