% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExamplePlot}
\alias{ExamplePlot}
\title{The EVA plot used to form the example scenarios}
\format{
A data frame with 6 rows and 3 columns, the definitions of which are:
\describe{
\item{taxon_name}{The scientific names of the taxa present in the example plot.}
\item{taxon_code}{The taxon, see \code{elements::ModellingTaxaLookup}.}
\item{cover_perc}{The percentage cover of the taxa present in example plot.}
}
}
\usage{
ExamplePlot
}
\description{
A dataset containing the taxon presences and percentage cover for the EVA plot used to form the example scenarios present in \code{elements::ExampleScenarios}.
}
\details{
\code{ExamplePlot}
}
\keyword{datasets}
