% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleScenarios}
\alias{ExampleScenarios}
\title{Predictor data for three example scenarios}
\format{
A data frame with 36 rows and 16 columns, the definitions of which are:
\describe{
\item{L}{Light}
\item{M}{Soil Moisture}
\item{N}{Soil Nitrogen}
\item{R}{Reaction}
\item{S}{Salinity}
\item{SD}{Soil Disturbance}
\item{GP}{Grazing Pressure}
\item{tmax_sm}{Maximum temperature in the warmest month}
\item{tmin_wt}{Minimum temperature in the coldest month}
\item{prec_wt}{Precipitation in the wettest quarter}
\item{prec_sm}{Precipitation in the driest quarter}
\item{scenario_code}{The scenario code.}
\item{scenario_stage}{The scenario stage, either 1) baseline, or 2) projection.}
\item{climate_scenario}{The SSP climate scenario.}
\item{timeslice}{The scenario timeslice, either a year or period.}
\item{scenario_name}{The scenario name.}
}
}
\usage{
ExampleScenarios
}
\description{
A dataset containing predictor data for five example scenarios, in three groups:
(A) Climate Change - RCP4.5 and (A) Climate Change - RCP8.5,
(B) Grazing Intensification and (B) Grazing Reduction, abd
(C) Nutrient Enrichment.
}
\details{
\code{ExampleScenarios}
}
\keyword{datasets}
