% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ExampleScenarios}
\alias{ExampleScenarios}
\title{Predictor data for three example scenarios}
\format{
A data frame with 26 rows and 14 columns, the definitions of which are:
\describe{
\item{L}{Light}
\item{M}{Soil Moisture}
\item{N}{Soil Nitrogen}
\item{R}{Reaction}
\item{S}{Salinity}
\item{SD}{Soil Disturbance}
\item{GP}{Grazing Pressure}
\item{bio05}{Maximum temperature in the warmest month}
\item{bio06}{Minimum temperature in the coldest month}
\item{bio16}{Precipitation in the wettest quarter}
\item{bio17}{Precipitation in the driest quarter}
\item{scenario}{The scenario name.}
\item{timeslice}{The scenario timeslice, either a year or period.}
\item{scenario_code}{The scenario code: a, b, or c.}
}
}
\usage{
ExampleScenarios
}
\description{
A dataset containing predictor data for five example scenarios, in three groups:
(A) Climate Change - RCP4.5 and (A) Climate Change - RCP8.5,
(B) Grazing Intensification and (B) Grazing Reduction, abd
(C) Nutrient Enrichment.
}
\details{
\code{ExampleScenarios}
}
\keyword{datasets}
