% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{Imbalances}
\alias{Imbalances}
\title{Presence-Absence Imbalances}
\format{
A data frame with 28683 rows and 6 columns, the definitions of which are:
\describe{
\item{Stage}{The plot selection 'stage'.}
\item{Absent}{The number of absences.}
\item{Present}{The number or presences.}
\item{PA_Imbalance}{The presence-absence imbalance ratio.}
\item{AP_Imbalance}{The absence-presence imbalance ratio.}
\item{taxon_code}{The taxon, see \code{elements::ModelledTaxaCodes}.}
}
}
\usage{
Imbalances
}
\description{
The number of presences and absences in the EVA, suitable samples within the EVA, and the training and test data, along with the presence-absence and absence-presence imbalance ratios.
}
\details{
\code{Imbalances}
}
\keyword{datasets}
