% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModelledTaxaCodes}
\alias{ModelledTaxaCodes}
\title{Taxon codes for taxa with models}
\format{
A vector containing  strings.
}
\usage{
ModelledTaxaCodes
}
\description{
Taxon codes for taxa which were successfully modelled as present in the taxon_code
column of \code{elements::ModellingTaxaLookup}.
}
\details{
\code{ModelledTaxaCodes}
}
\keyword{datasets}
