% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModellingTaxaLookup}
\alias{ModellingTaxaLookup}
\title{A lookup between the taxon names, taxon codes and success of selected workflow steps}
\format{
A data frame with 11228 rows and 8 columns, the definitions of which are:
\describe{
\item{taxon_name}{Taxon names.}
\item{taxon_code}{The taxon code, see \code{elements::TaxonomicBackbone}}
\item{imbalance}{A boolean (TRUE/FALSE) representing whether ... were successfully calculated, see \code{elements::Imbalances}.}
\item{nichewidth}{A boolean (TRUE/FALSE) representing whether ... were successfully calculated, see \code{elements::NicheWidths}.}
\item{performance}{A boolean (TRUE/FALSE) representing whether ... were successfully calculated, see \code{elements::PerformanceMeasures}.}
\item{ale}{A boolean (TRUE/FALSE) representing whether ... were successfully calculated, see \code{elements::ALEData}.}
\item{pdp}{A boolean (TRUE/FALSE) representing whether ... were successfully calculated, see \code{elements::PDPData}.}
\item{model}{A boolean (TRUE/FALSE) representing whether ... were successfully fitted.}
}
}
\usage{
ModellingTaxaLookup
}
\description{
A lookup between the taxon names, taxon codes and success of selected workflow steps
in the model-fitting process.
}
\details{
\code{ModellingTaxaLookup}
}
\keyword{datasets}
