% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{ModelObjectAvailability}
\alias{ModelObjectAvailability}
\title{Model Object Availability}
\format{
A data frame with 8897 rows and 8 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon, see \code{elements::ModelledTaxaCodes}.}
\item{imbalance}{A boolean value indicating whether the taxon has data in \code{elements::Imbalances}.}
\item{nichewidth}{A boolean value indicating whether the taxon has data in \code{elements::NicheWidths}.}
\item{performance}{A boolean value indicating whether the taxon has data in \code{elements::PerformanceMeasures}.}
\item{ale}{A boolean value indicating whether the taxon has data in \code{elements::ALEData}.}
\item{pdp}{A boolean value indicating whether the taxon has data in \code{elements::PDPData}.}
\item{hoa}{A boolean value indicating whether the taxon has data in \code{elements::HOAData}.}
\item{model}{A boolean value indicating whether the taxon has a ecological niche model as present in ./inst/extdata/Models.}
}
}
\usage{
ModelObjectAvailability
}
\description{
A data frame containing boolean (TRUE/FALSE) values indicating the objects/information
available for each modelled taxon.
}
\details{
\code{ALEDaModelObjectAvailabilityta}
}
\keyword{datasets}
