% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{PDPData}
\alias{PDPData}
\title{PDP data}
\format{
A data frame with 1351900 rows and 4 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon, see \code{elements::ModelTaxa}.}
\item{x}{The variable value.}
\item{y}{The PDP value.}
\item{variable}{The variable name.}
}
}
\usage{
PDPData
}
\description{
A dataset containing the Partial Dependency Profile (PDP) (Molnar, 2018) data for the models, generated using 1000 samples from the training data.
}
\details{
\code{PDPData}
}
\keyword{datasets}
