% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{performanceMeasures}
\alias{performanceMeasures}
\title{Model performance measures}
\format{
A data frame with 4 rows and 11 columns, the definitions of which are:
\describe{
\item{species}{The species, represented by a UKSI TVK.}
\item{bbrier}{Binary Brier score}
\item{logloss}{...}
\item{auc}{Area under the ROC curve}
\item{prauc}{Area under the precision-recall curve}
\item{precision}{Positive predictive value}
\item{recall}{...}
\item{sensitivity}{True positive rate}
\item{specificity}{True negative rate}
\item{acc}{Accuracy}
\item{bacc}{Balanced accuracy}
}
}
\usage{
performanceMeasures
}
\description{
A dataset containing a number of performance metrics for each model.
}
\details{
\code{performanceMeasures}
}
\keyword{datasets}
