% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_be.R
\name{plot_be}
\alias{plot_be}
\title{Plot the bivariate effects for a taxon}
\usage{
plot_be(taxon, vars)
}
\arguments{
\item{taxon}{A string containing a taxon name as present in \code{elements::ModelledTaxaCodes}.}

\item{vars}{A vector of strings containing two variable names as present in \code{elements::VariableNames}}
}
\value{
A plot containing the bivariate probability distribution of a taxon for two variables.
}
\description{
Generate a plot containing the bivariate probability distribution of a taxon for two variables
generated using the Hold-At-Optima (HOA) method present in \code{elements::predict_occ}.
}
\examples{
elements::startup(); elements::plot_be(taxon = "carex_rostrata", vars = c("tmax_sm", "tmin_wt"))
}
