% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me.R
\name{plot_me}
\alias{plot_me}
\title{Plot the marginal effect values for a taxon and set of variables}
\usage{
plot_me(taxon, me_type, free_y, presences, eivs, vars)
}
\arguments{
\item{taxon}{The taxon_code, see \code{elements::TaxaBackbone}.}

\item{me_type}{A string representing the marginal effect plot type, one of "ale" or "pdp".}

\item{free_y}{A boolean. If TRUE the Y axis scales are independent and free for all subplots. If FALSE the Y axis scales are fixed between all subplots.}

\item{presences}{A boolean. If TRUE a box and whiskers plot showing the distribution of presences along each variable will be displayed.}

\item{eivs}{A boolean. If TRUE a point representing the EIV value and arrows representing the EIV niche widths for the taxon will be displayed, where available in \code{elements::VariableData}.}

\item{vars}{A vector of variables. Must include atleast one of the following columns: "L", "M", "N", "R", "S", "SD", "GP", "bio05", "bio06", "bio16", and "bio17".}
}
\value{
A composite plot showing the marginal effects and optionally the distribution of presences for selected model variables.
}
\description{
Plot the Accumulated Local Effect (ALE) or Partial Dependence Profile (PDP)
marginal effect/univariate response curves for a selected taxon and variables.
}
\details{
The presence-absence imbalance in the training data varies by taxon.
This 'ghost of imbalance' (Jiménez-Valverde & Lobo, 2006)
has several impacts on the PDP plots:
\enumerate{
\item the optimum value of the PDP curve may be less than 1.
\item the entire PDP curve may sit below y = 0.5 (the presence-absence threshold).
}

When inspecting the PDP plots, it is therefore important to pay more attention
to the shape of the response, rather than the absolute PDP value.
}
\examples{
elements::plot_me(taxon = "stellaria_graminea", me_type = "ale", free_y = FALSE, presences = TRUE, eivs = TRUE, vars = c("L", "M", "N", "R", "S", "SD", "GP", "bio05", "bio06", "bio16", "bio17"))

}
\references{
Jiménez-Valverde, A., Lobo, J.M., 2006. The ghost of unbalanced species distribution data in geographical model predictions. Diversity and Distributions 12, 521–524. https://doi.org/10.1111/j.1366-9516.2006.00267.x
}
