% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_me.R
\name{plot_me}
\alias{plot_me}
\title{Plot the marginal effect values for a taxon and set of variables}
\usage{
plot_me(taxon, me_type, free_y, presences, vars)
}
\arguments{
\item{taxon}{The taxon_code, see \code{elements::ModellingSpecies} and \code{elements::TaxaBackbone}.}

\item{me_type}{A string representing the marginal effect plot type, one of "ale" or "pdp".}

\item{free_y}{A boolean. If TRUE the Y axis scales are independent and free for all subplots. If FALSE the Y axis scales are fixed between all subplots.}

\item{presences}{A boolean. If true a box and whiskers plot showing the distribution of presences along each variable will be displayed.}

\item{vars}{A vector of variables. Must include atleast one of the following columns: "L", "M", "N", "R", "S", "SD", "GP", "bio05", "bio06", "bio16", and "bio17".}
}
\value{
A composite plot showing the ALE effects for each model variable
}
\description{
Plot the Accumulated Local Effect (ALE) or Partial Dependence Profile (PDP)
marginal effect values for a selected taxon and variables.
}
\examples{
elements::plot_me(taxon = "stellaria_graminea", me_type = "ale", free_y = FALSE, presences = TRUE, vars = c("L", "M", "N", "R", "S", "SD", "GP", "bio05", "bio06", "bio16", "bio17"))
}
