% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occ.R
\name{predict_occ}
\alias{predict_occ}
\title{Generate predictions for the probability of occurrence for specified taxa}
\usage{
predict_occ(
  taxa,
  predictors,
  pa = "Present",
  limit = NULL,
  holdopt = NULL,
  dp = 3,
  append = "ids"
)
}
\arguments{
\item{taxa}{A vector of strings containing one or more taxa to generate predictions for. See \code{elements::ModelledTaxaCodes}. Optional.}

\item{predictors}{A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, tmax_sm, tmin_wt, prec_wt, prec_sm, and optionally taxon}

\item{pa}{One of "Present", "Absent", or c("Present", "Absent").}

\item{limit}{A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q25_q75". Which if set assigns a probability of 0 to a set of predictors if one or more of those predictors are outside the stipulated quantile ranges. Only applied if pa = "Present". Optional.}

\item{holdopt}{Hold one or more variables at their optimum values. NULL by default, else a vector of variable codes, e.g. c("SD", "GP").}

\item{dp}{The number of decimal places to round the probability values to.}

\item{append}{A string, one of "all", "predictors", or "ids" representing which columns from the predictors data frame to return with the results.}
}
\value{
A data frame containing the probability of occurrence (Present and/or Absent) for each taxon.
}
\description{
Generate predictions for the probability of occurrence (presence and/or absence) using a data frame of predictors for a one or more taxa specified in either:
\enumerate{
\item the 'taxa' argument, or
\item a column in the predictors data frame named 'taxon_code'
}
}
\details{
A number of optional arguments provide additional control over the use of the models.
First, using the 'limit' argument the probabilities may be set to 0 if one or
more of the predictor values are outside a stipulated quantile range,
e.g. the 1\% and 99\% quantiles by setting the 'limit' argument to "q01_q99".
This may be used to strictly enforce the assignment of probability values of 0.
As the quantile values present in \code{elements::NicheWidths} are derived using all presences of each taxon in the EVA,
they may better represent the upper and lower tolerances of a taxons ecological niche,
especially for taxa where the presences were undersampled.
Second, using the 'holdopt' argument one or more predictor variables can be held constant,
e.g. to hold soil disturbance constant supply c("SD") to the 'holdopt' argument.
This may be useful in instances where you wish to remove the influence of one or more variables
on the model results.

If taxon codes are supplied in the 'taxa' argument and there is a column in the predictors data frame named 'taxon_code',
the taxa present in the 'taxa' argument will be used and results will be calculated using the entire predictors data frame
and the 'taxon_code' column will be replaced.

NOTE: to use this function you must first run \code{elements::startup()}
}
\examples{
\dontrun{
elements::startup()

# Generate predictions using a data frame containing taxon_codes in the 'taxon' column.
elements::predict_occ(taxa = NULL, predictors = elements::ExampleData2, pa = "Present", limit = NULL, holdopt = c("SD", "GP"))

# Generate predictions for taxa specified in the 'taxa' argument using a data frame containing only predictor variables.
elements::predict_occ(taxa = c("stellaria_graminea", "silene_flos-cuculi_aggr"), predictors = elements::ExampleData1, pa = "Present", limit = NULL, holdopt = c("SD", "GP"))

# Generate predictions for taxa specified in the 'taxa' argument using a data frame containing both predictor variables and a 'taxon_code' column (which is ignored and overwritten).
elements::predict_occ(taxa = c("stellaria_graminea", "silene_flos-cuculi_aggr"), predictors = elements::ExampleData2, pa = "Present", limit = NULL, holdopt = c("SD", "GP"))

elements::shutdown() 
}
}
