% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occ.R
\name{predict_occ}
\alias{predict_occ}
\title{Generate predictions for the probability of occurrence}
\usage{
predict_occ(taxon, predictors, pa = "Present", dp = 3)
}
\arguments{
\item{taxon}{The taxon_code, see \code{elements::ModellingSpecies} and \code{elements::TaxaBackbone}.}

\item{predictors}{A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, and bio17}

\item{pa}{One of "Present", "Absent", or c("Present", "Absent").}

\item{dp}{The number of decimal places to round the probability values to.}
}
\value{
A data frame containing the probability of occurrence.
}
\description{
Generate predictions for the probability of occurrence (presence and/or absence) for a given taxon.
}
\details{
NOTE: to use this function you must first run \code{elements::startup()}
}
\examples{
\dontrun{
elements::startup()
elements::predict_occ(taxon = "stellaria_graminea",  predictors = elements::ExampleData, pa = "Present")
elements::shutdown() 
}
}
