% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/predict_occ_taxon.R
\name{predict_occ_taxon}
\alias{predict_occ_taxon}
\title{Generate predictions for the probability of occurrence}
\usage{
predict_occ_taxon(
  taxon,
  predictors,
  pa = "Present",
  limit = NULL,
  dp = 3,
  append_predictors = TRUE
)
}
\arguments{
\item{taxon}{The taxon_code, see \code{elements::TaxaBackbone}.}

\item{predictors}{A data frame of predictors. Must include the following columns: L, M, N, R, S, SD, GP, bio05, bio06, bio16, and bio17}

\item{pa}{One of "Present", "Absent", or c("Present", "Absent").}

\item{limit}{A string representing the niche width quantiles, one of "min_max", "q01_q99", "q05_q95", "q25_q75". Which if set assigns a probability of 0 to the Present column and/or 1 to the Absent column to a set of predictors if one or more of those predictors are outside the stipulated quantile ranges. Optional.}

\item{dp}{The number of decimal places to round the probability values to.}

\item{append_predictors}{A boolean. If TRUE return the predictors data frame with the results in an additional column.}
}
\value{
A data frame containing the probability of occurrence (Present and/or Absent), if specified, appended to the predictors data.
}
\description{
Generate predictions for the probability of occurrence (presence and/or absence) for a given taxon.
}
\details{
NOTE: to use this function you must first run \code{elements::startup()}
}
\examples{
\dontrun{
elements::startup()
elements::predict_occ_taxon(taxon = "stellaria_graminea", predictors = elements::ExampleData1, pa = "Present", limit = NULL, dp = 3, append_predictors = TRUE)
elements::shutdown() 
}
}
