% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/startup.R
\name{startup}
\alias{startup}
\title{Initiliase a connection to the filehash database containing the Ecological Niche Models (ENMs)}
\usage{
startup(models = NULL)
}
\arguments{
\item{models}{One of "all", "test", or NULL, indicating whether the complete set of models should be loaded, or the test set of models. NULL by default, which loads all models.}
}
\value{
A connection to the filehash database in the global environment containing the ENMs
}
\description{
Initiliase a connection to the filehash database containing the Ecological Niche Models (ENMs)
}
\examples{
elements::startup()
}
