% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxonomicBackbone}
\alias{TaxonomicBackbone}
\title{Taxonomic backbone}
\format{
A data frame with 18652 rows and 9 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon, see \code{elements::ModelledTaxaCodes}.}
\item{taxon_code_no_suffix}{The taxon code with no 'aggr' suffix.}
\item{Root}{The Root taxon.}
\item{Phylum}{The parent Phylum taxon associated with the Species.}
\item{Subdivision}{The parent Subdivision taxon associated with the Species.}
\item{Class}{The parent Class taxon associated with the Species.}
\item{Superorder}{The parent Superorder taxon associated with the Species.}
\item{Family}{The parent Family taxon associated with the Species.}
\item{Genus}{The parent Genus taxon associated with the Species.}
}
}
\usage{
TaxonomicBackbone
}
\description{
The higher taxa associated with the modelled taxa, retrieved from EuroSL.
}
\details{
\code{TaxonomicBackbone}
}
\keyword{datasets}
