% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{TaxonomicBackbone}
\alias{TaxonomicBackbone}
\title{Taxonomic backbone}
\format{
A data frame with 6177 rows and 25 columns, the definitions of which are:
\describe{
\item{taxon_code}{The taxon codes used throughout the package, formed from the taxon_name values by coercing all letters to lower and replacing whitespace with underscores.}
\item{taxon_name}{The taxon names for the modelled taxa, as they appear in the EVA.}
\item{taxon_concept}{The acccepted taxon concepts for each taxon_name value, following the authority of the EuroSL - see \code{elements::EuroSL}}
\item{eurosl_status}{The taxonomic status of the taxon_name, following the authority of the EuroSL.}
\item{eurosl_id}{The EuroSL ID code for the taxon_name.}
\item{Species}{The Species name associated with the taxon_name.}
\item{Genus}{The parent Genus taxon associated with the taxon_name.}
\item{Tribe}{The parent Tribe taxon associated with the taxon_name.}
\item{Family}{The parent Family taxon associated with the taxon_name.}
\item{Order}{The parent Order taxon associated with the taxon_name.}
\item{Superorder}{The parent Superorder taxon associated with the taxon_name.}
\item{Class}{The parent Class taxon associated with the taxon_name.}
\item{Subdivision}{The parent Subdivision taxon associated with the taxon_name.}
\item{Phylum}{The parent Phylum taxon associated with the taxon_name.}
\item{Root}{The parent Root taxon associated with the taxon_name.}
\item{Section}{The Section name associated with the taxon_name.}
\item{Species Aggregate}{The Species Aggregate name associated with the taxon_name.}
\item{Subclass}{The parent Subclass taxon associated with the taxon_name.}
\item{Division}{The parent Division taxon associated with the taxon_name.}
\item{Suprageneric Taxon}{The parent Suprageneric Taxon taxon associated with the taxon_name.}
\item{Subfamily}{The parent Subfamily taxon associated with the taxon_name.}
\item{Subspecies}{The Subspecies name associated with the taxon_name.}
\item{Variety}{The Variety name associated with the taxon_name.}
\item{Coll. species}{The Coll. species name associated with the taxon_name.}
\item{n_presences_eva}{The number of presences of the taxon (taxon_name) in the EVA, only considering plots with valid coordinates and years.}
}
}
\usage{
TaxonomicBackbone
}
\description{
The taxon names and codes for the modeled taxa in the EVA, with the associated taxon concept information from EuroSL, and parent taxa.
}
\details{
\code{TaxonomicBackbone}
}
\keyword{datasets}
