% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VariableData}
\alias{VariableData}
\title{Environmental indicator value data}
\format{
A data frame with 9406 rows and 27 columns, the definitions of which are:
\describe{
\item{taxon_name}{The taxon name.}
\item{taxon_code}{The taxon, see \code{elements::ModelledTaxaCodes}. Not all modelled taxa have EIV values from the original datasets.}
\item{DF}{Disturbance frequency (Midolo et al., 2023)}
\item{DF.sd}{Disturbance frequency standard deviation (Midolo et al., 2023)}
\item{DFH.sd}{Disturbance frequency herb layer standard deviation (Midolo et al., 2023)}
\item{DFh}{Disturbance frequency herb layer (Midolo et al., 2023)}
\item{DS}{Disturbance severity (Midolo et al., 2023)}
\item{DS.sd}{Disturbance severity standard deviation (Midolo et al., 2023)}
\item{DSH.sd}{Disturbance severity herb layer standard deviation (Midolo et al., 2023)}
\item{DSh}{Disturbance severity herb layer (Midolo et al., 2023)}
\item{GP}{Grazing pressure (Midolo et al., 2023)}
\item{GP.sd}{Grazing pressure standard deviation (Midolo et al., 2023)}
\item{L}{Light (Dengler et al., 2023)}
\item{L.nw}{Light niche width (Dengler et al., 2023)}
\item{M}{Moisture (Dengler et al., 2023)}
\item{M.nw}{Moisture niche width (Dengler et al., 2023)}
\item{MF}{Mowing frequency (Midolo et al., 2023)}
\item{MF.sd}{Mowing frequency standard deviation (Midolo et al., 2023)}
\item{N}{Nitrogen (Dengler et al., 2023)}
\item{N.nw}{Nitrogen niche width (Dengler et al., 2023)}
\item{R}{Reaction (Dengler et al., 2023)}
\item{R.nw}{Reaction niche width (Dengler et al., 2023)}
\item{S}{Salinity (Tichý et al 2023)}
\item{SD}{Soil disturbance (Midolo et al., 2023)}
\item{SD.sd}{Soil disturbance standard deviation (Midolo et al., 2023)}
\item{T}{Temperature (Dengler et al., 2023)}
\item{T.nw}{Temperature niche width (Dengler et al., 2023)}
}
}
\usage{
VariableData
}
\description{
A dataset containing a combined dataset of Environmental Indicator Values (EIVs) from Dengler et al (2023), Midolo et al (2023), and Tichý et al (2023).
}
\details{
\code{VariableData}
}
\references{
{
Marchi, M., Castellanos-Acuña, D., Hamann, A., Wang, T., Ray, D., Menzel, A., 2020. ClimateEU, scale-free climate normals, historical time series, and future projections for Europe. Sci Data 7, 428. https://doi.org/10.1038/s41597-020-00763-0

Dengler, J., Jansen, F., Chusova, O., Hüllbusch, E., Nobis, M.P., Meerbeek, K.V., Axmanová, I., Bruun, H.H., Chytrý, M., Guarino, R., Karrer, G., Moeys, K., Raus, T., Steinbauer, M.J., Tichý, L., Tyler, T., Batsatsashvili, K., Bita-Nicolae, C., Didukh, Y., Diekmann, M., Englisch, T., Fernández-Pascual, E., Frank, D., Graf, U., Hájek, M., Jelaska, S.D., Jiménez-Alfaro, B., Julve, P., Nakhutsrishvili, G., Ozinga, W.A., Ruprecht, E.-K., Šilc, U., Theurillat, J.-P., Gillet, F., 2023. Ecological Indicator Values for Europe (EIVE) 1.0. Vegetation Classification and Survey 4, 7–29. https://doi.org/10.3897/VCS.98324

Midolo, G., Herben, T., Axmanová, I., Marcenò, C., Pätsch, R., Bruelheide, H., Karger, D.N., Aćić, S., Bergamini, A., Bergmeier, E., Biurrun, I., Bonari, G., Čarni, A., Chiarucci, A., De Sanctis, M., Demina, O., Dengler, J., Dziuba, T., Fanelli, G., Garbolino, E., Giusso del Galdo, G., Goral, F., Güler, B., Hinojos-Mendoza, G., Jansen, F., Jiménez-Alfaro, B., Lengyel, A., Lenoir, J., Pérez-Haase, A., Pielech, R., Prokhorov, V., Rašomavičius, V., Ruprecht, E., Rūsiņa, S., Šilc, U., Škvorc, Ž., Stančić, Z., Tatarenko, I., Chytrý, M., 2023. Disturbance indicator values for European plants. Global Ecology and Biogeography 32, 24–34. https://doi.org/10.1111/geb.13603

Tichý, L., Axmanová, I., Dengler, J., Guarino, R., Jansen, F., Midolo, G., Nobis, M.P., Van Meerbeek, K., Aćić, S., Attorre, F., Bergmeier, E., Biurrun, I., Bonari, G., Bruelheide, H., Campos, J.A., Čarni, A., Chiarucci, A., Ćuk, M., Ćušterevska, R., Didukh, Y., Dítě, D., Dítě, Z., Dziuba, T., Fanelli, G., Fernández-Pascual, E., Garbolino, E., Gavilán, R.G., Gégout, J.-C., Graf, U., Güler, B., Hájek, M., Hennekens, S.M., Jandt, U., Jašková, A., Jiménez-Alfaro, B., Julve, P., Kambach, S., Karger, D.N., Karrer, G., Kavgacı, A., Knollová, I., Kuzemko, A., Küzmič, F., Landucci, F., Lengyel, A., Lenoir, J., Marcenò, C., Moeslund, J.E., Novák, P., Pérez-Haase, A., Peterka, T., Pielech, R., Pignatti, A., Rašomavičius, V., Rūsiņa, S., Saatkamp, A., Šilc, U., Škvorc, Ž., Theurillat, J.-P., Wohlgemuth, T., Chytrý, M., 2023. Ellenberg-type indicator values for European vascular plant species. Journal of Vegetation Science 34, e13168. https://doi.org/10.1111/jvs.13168
}
}
\keyword{datasets}
