% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{VariableNames}
\alias{VariableNames}
\title{Model variables codes}
\format{
A vector containing  strings.
}
\usage{
VariableNames
}
\description{
A vector containing the environmental variable codes for the model variables;
equivalent to the variable_code values in the \code{elements::VariableLookup} data frame for rows where model_var is TRUE.
}
\details{
\code{VariableNames}
}
\keyword{datasets}
